+++ R/properties.R
@@ -33,7 +33,28 @@ properties <- list(
     creator     = c("creator", "login",        as = "character"),
     created_at  = c("created_at",              as = "datetime"),
     updated_at  = c("updated_at",              as = "datetime"),
-    content_url = c("content_url",             as = "character"))
+    content_url = c("content_url",             as = "character")),
+
+  users = list(
+    id          = c("id",                      as = "integer"),
+    login       = c("login",                   as = "character"),
+    type        = c("type",                    as = "character"),
+    site_admin  = c("site_admin",              as = "logical"),
+    html_url    = c("html_url",                as = "character")),
+
+  user = list(
+    id          = c("id",                      as = "integer"),
+    login       = c("login",                   as = "character"),
+    name        = c("name",                    as = "character"),
+    email       = c("email",                   as = "character"),
+    blog        = c("blog",                    as = "character"),
+    company     = c("company",                 as = "character"),
+    location    = c("location",                as = "character"),
+    hireable    = c("hireable",                as = "logical"),
+    bio         = c("bio",                     as = "character"),
+    type        = c("type",                    as = "character"),
+    site_admin  = c("site_admin",              as = "logical"),
+    html_url    = c("html_url",                as = "character"))
 
 )
 
@@ -57,6 +78,9 @@ values <- list(
   card = list(
     content_type   = c("Issue", "PullRequest"),
     archived_state = c("archived", "not_archived", "all"),
-    position       = c("top", "bottom"))
+    position       = c("top", "bottom")),
+
+  user = list(
+    role           = c("admin", "member", "all"))
 
 )